% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DEGAS_utils.R
\name{centerFunc}
\alias{centerFunc}
\title{Center DEGAS scores}
\usage{
centerFunc(x, convert = TRUE)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{convert}{Whether to convert Python result to R.}
}
\value{
Centered vector or matrix.
}
\description{
Center DEGAS scores
}
