% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-python-loader.R
\name{degas_py}
\alias{degas_py}
\title{Load bundled Python DEGAS module}
\usage{
degas_py(convert = TRUE, reload = FALSE)
}
\arguments{
\item{convert}{Whether reticulate should convert Python objects to R objects.}

\item{reload}{Force re-importing the Python module.}
}
\value{
Python module object.
}
\description{
This function lazily imports the Python module under `inst/DEGAS`.
Users normally do not need to call this directly.
}
