% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-python-loader.R
\name{degas_py_available}
\alias{degas_py_available}
\title{Check whether bundled Python DEGAS backend can be loaded}
\usage{
degas_py_available()
}
\value{
TRUE/FALSE.
}
\description{
Check whether bundled Python DEGAS backend can be loaded
}
