% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DEGAS_utils.R
\name{knnSmooth}
\alias{knnSmooth}
\title{kNN smooth DEGAS scores}
\usage{
knnSmooth(probs, locs, k = 5L, convert = TRUE)
}
\arguments{
\item{probs}{DEGAS scores, vector or matrix.}

\item{locs}{Coordinates, for example UMAP or tSNE matrix.}

\item{k}{Number of nearest neighbors.}

\item{convert}{Whether to convert Python result to R.}
}
\value{
Smoothed scores.
}
\description{
kNN smooth DEGAS scores
}
