% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23-predClass.R
\name{predClass}
\alias{predClass}
\title{Make Predictions Using Trained DEGAS Model}
\usage{
predClass(ccModel1, Exp, scORpat)
}
\arguments{
\item{ccModel1}{A trained DEGAS model object}

\item{Exp}{Expression matrix for prediction}

\item{scORpat}{Indicator for single-cell or patient data}
}
\value{
Model predictions
}
\description{
Makes class predictions based on a trained DEGAS model. Automatically
routes to the appropriate prediction function based on the model architecture.
}
\examples{
\dontrun{
# Make predictions using a trained model
predictions <- predClass(trained_model, expression_matrix, "sc")
}

}
