% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23-predClass.R
\name{predClass2}
\alias{predClass2}
\title{Prediction from DenseNet Architecture Model}
\usage{
predClass2(ccModel1, Exp, scORpat)
}
\arguments{
\item{ccModel1}{A trained DEGAS model object}

\item{Exp}{Expression matrix for prediction}

\item{scORpat}{Indicator for single-cell ("sc") or patient ("pat") data}
}
\value{
Model predictions
}
\description{
Makes predictions using a trained DEGAS model with DenseNet architecture.
}
