% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DEGAS_api.R
\name{predict_degas}
\alias{predict_degas}
\title{Predict cell-level DEGAS scores}
\usage{
predict_degas(
  model,
  sc_matrix,
  sc_gene_names,
  cell_names = NULL,
  batch_size = 2048L,
  convert = TRUE
)
}
\arguments{
\item{model}{A Python DEGASTensorFlow model returned by train_degas.}

\item{sc_matrix}{Single-cell expression matrix, cells x genes.}

\item{sc_gene_names}{Gene names for columns of sc_matrix.}

\item{cell_names}{Optional cell names.}

\item{batch_size}{Prediction batch size.}

\item{convert}{Whether to convert Python pandas.DataFrame to R data.frame.}
}
\value{
data.frame if convert=TRUE, otherwise Python pandas.DataFrame.
}
\description{
Predict cell-level DEGAS scores
}
