% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DEGAS_utils.R
\name{preprocessCounts}
\alias{preprocessCounts}
\title{Preprocess count matrix using DEGAS-style transform}
\usage{
preprocessCounts(X, convert = TRUE)
}
\arguments{
\item{X}{Count matrix, genes x samples/cells.}

\item{convert}{Whether to convert Python result to R.}
}
\value{
Processed matrix, samples/cells x genes.
}
\description{
Preprocess count matrix using DEGAS-style transform
}
