% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DEGAS_utils.R
\name{toCorrCoeff}
\alias{toCorrCoeff}
\title{Convert DEGAS probabilities to association scores}
\usage{
toCorrCoeff(probs, convert = TRUE)
}
\arguments{
\item{probs}{Numeric vector or matrix.}

\item{convert}{Whether to convert Python result to R.}
}
\value{
Numeric vector or matrix.
}
\description{
Convert DEGAS probabilities to association scores
}
