% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22-makeExec.R
\name{makeExec}
\alias{makeExec}
\title{Generate Python Executable for Neural Network}
\usage{
makeExec(
  FFdepth,
  model_type = c("ClassClass", "ClassCox", "ClassBlank", "BlankClass", "BlankCox"),
  DEGAS.toolsPath = system.file("DEGAS_tools", package = "DEGAS", mustWork = TRUE),
  architecture = c("Standard", "DenseNet"),
  ...
)
}
\arguments{
\item{FFdepth}{Number of hidden layers in the feedforward network}

\item{model_type}{Type of neural network model. Must be one of:
'BlankClass', 'ClassBlank', 'BlankCox', 'ClassClass', or 'ClassCox'}

\item{tmpDir}{Temporary directory for script generation}
}
\value{
None. Generates Python scripts in the specified temporary directory.
}
\description{
Creates a Python executable for feedforward neural network implementation
by combining multiple script parts into a complete TensorFlow model.
}
\examples{
\dontrun{
# Create a 2-layer classifier model
makeExec("/tmp/model/", 2, "ClassClass")
}

}
