% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22-makeExec.R
\name{makeExec2}
\alias{makeExec2}
\title{Generate DenseNet Python Executable}
\usage{
makeExec2(tmpDir, FFdepth, model_type)
}
\arguments{
\item{tmpDir}{Temporary directory for script generation}

\item{FFdepth}{Number of hidden layers in the DenseNet}

\item{model_type}{Type of neural network model. Must be one of:
'BlankClass', 'ClassBlank', 'BlankCox', 'ClassClass', or 'ClassCox'}
}
\value{
None. Generates Python scripts in the specified temporary directory.
}
\description{
Creates a Python executable for DenseNet (densely connected neural network)
implementation by combining multiple script parts into a complete TensorFlow model.
}
\examples{
\dontrun{
# Create a 3-layer DenseNet classifier
makeExec2("/tmp/densenet/", 3, "ClassClass")
}

}
