% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-runCCMTLBag.R
\name{runCCMTLBag.optimized}
\alias{runCCMTLBag.optimized}
\title{Optimized Bootstrap Aggregation for Cross-Condition Multi-Task Learning}
\usage{
runCCMTLBag.optimized(
  verbose = SigBridgeRUtils::getFuncOption("verbose") \%||\% TRUE,
  ...
)
}
\description{
Optimized Bootstrap Aggregation for Cross-Condition Multi-Task Learning
}
\references{
Johnson TS, Yu CY, Huang Z, Xu S, Wang T, Dong C, et al. Diagnostic Evidence GAuge of Single cells (DEGAS): a flexible deep transfer learning framework for prioritizing cells in relation to disease. Genome Med. 2022 Feb 1;14(1):11.
}
\seealso{
Other DEGAS: 
\code{\link{LabelBinaryCells}()},
\code{\link{LabelContinuousCells}()},
\code{\link{LabelSurvivalCells}()},
\code{\link{Vec2sparse}()},
\code{\link{predClassBag.optimized}()},
\code{\link{readOutputFiles.optimized}()},
\code{\link{runCCMTL.optimized}()},
\code{\link{writeInputFiles.optimized}()}
}
\concept{DEGAS}
